+++ src/test/java/com/woowacamp/soolsool/acceptance/core/member/MemberAcceptanceTest.java
@@ -1,7 +1,8 @@
-package com.woowacamp.soolsool.acceptance;
+package com.woowacamp.soolsool.acceptance.core.member;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.woowacamp.soolsool.acceptance.AcceptanceTest;
 import com.woowacamp.soolsool.core.member.dto.request.MemberAddRequest;
 import com.woowacamp.soolsool.core.member.dto.request.MemberMileageChargeRequest;
 import com.woowacamp.soolsool.core.member.dto.request.MemberModifyRequest;
@@ -29,17 +30,17 @@ class MemberAcceptanceTest extends AcceptanceTest {
         LoginRequest loginRequest = new LoginRequest(email, password);
 
         ExtractableResponse<Response> loginResponse = RestAssured
-                .given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(loginRequest).log().all()
-                .when().post("/auth/login")
-                .then().log().all()
-                .extract();
+            .given()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(loginRequest).log().all()
+            .when().post("/auth/login")
+            .then().log().all()
+            .extract();
 
         String accessToken = loginResponse.body().as(new TypeRef<ApiResponse<LoginResponse>>() {
-                })
-                .getData()
-                .getAccessToken();
+            })
+            .getData()
+            .getAccessToken();
         return accessToken;
     }
 
@@ -48,25 +49,25 @@ class MemberAcceptanceTest extends AcceptanceTest {
     void createMember() {
         // given
         MemberAddRequest memberAddRequest = new MemberAddRequest(
-                "CUSTOMER",
-                EMAIL,
-                PASSWORD,
-                "최배달",
-                "010-1234-5678",
-                "0",
-                "서울시 잠실역");
+            "CUSTOMER",
+            EMAIL,
+            PASSWORD,
+            "최배달",
+            "010-1234-5678",
+            "0",
+            "서울시 잠실역");
 
         // when
         ExtractableResponse<Response> response = RestAssured
-                .given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberAddRequest)
-                .log().all()
-                .when()
-                .post("/members")
-                .then()
-                .log().all()
-                .extract();
+            .given()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(memberAddRequest)
+            .log().all()
+            .when()
+            .post("/members")
+            .then()
+            .log().all()
+            .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
@@ -77,35 +78,35 @@ class MemberAcceptanceTest extends AcceptanceTest {
     void getMember() {
         // given
         MemberAddRequest memberAddRequest = new MemberAddRequest(
-                "CUSTOMER",
-                EMAIL,
-                PASSWORD,
-                "최배달",
-                "010-1234-5678",
-                "0",
-                "서울시 잠실역");
+            "CUSTOMER",
+            EMAIL,
+            PASSWORD,
+            "최배달",
+            "010-1234-5678",
+            "0",
+            "서울시 잠실역");
 
         RestAssured.given()
-                .when()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberAddRequest)
-                .post("/members")
-                .then();
+            .when()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(memberAddRequest)
+            .post("/members")
+            .then();
         String token = findToken(EMAIL, PASSWORD);
 
         // when
         ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-                .when().get("/members")
-                .then().log().all()
-                .extract();
+            .given().log().all()
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+            .when().get("/members")
+            .then().log().all()
+            .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
         MemberFindResponse memberFindResponse = response
-                .jsonPath()
-                .getObject("data", MemberFindResponse.class);
+            .jsonPath()
+            .getObject("data", MemberFindResponse.class);
         assertThat(memberFindResponse.getName()).isEqualTo("최배달");
     }
 
@@ -114,36 +115,36 @@ class MemberAcceptanceTest extends AcceptanceTest {
     void modifyMember() {
         // given
         MemberAddRequest memberAddRequest = new MemberAddRequest(
-                "CUSTOMER",
-                EMAIL,
-                PASSWORD,
-                "최배달",
-                "010-1234-5678",
-                "0",
-                "서울시 잠실역");
+            "CUSTOMER",
+            EMAIL,
+            PASSWORD,
+            "최배달",
+            "010-1234-5678",
+            "0",
+            "서울시 잠실역");
 
         RestAssured.given()
-                .when()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberAddRequest)
-                .post("/members")
-                .then();
+            .when()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(memberAddRequest)
+            .post("/members")
+            .then();
         String token = findToken(EMAIL, PASSWORD);
 
         MemberModifyRequest modifyRequest = new MemberModifyRequest(
-                "modify_password",
-                "modify_name",
-                "modify_address");
+            "modify_password",
+            "modify_name",
+            "modify_address");
 
         // when
         ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(modifyRequest)
-                .when().patch("/members")
-                .then().log().all()
-                .extract();
+            .given().log().all()
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(modifyRequest)
+            .when().patch("/members")
+            .then().log().all()
+            .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
@@ -154,29 +155,29 @@ class MemberAcceptanceTest extends AcceptanceTest {
     void deleteMember() {
         // given
         MemberAddRequest memberAddRequest = new MemberAddRequest(
-                "CUSTOMER",
-                EMAIL,
-                PASSWORD,
-                "최배달",
-                "010-1234-5678",
-                "0",
-                "서울시 잠실역");
+            "CUSTOMER",
+            EMAIL,
+            PASSWORD,
+            "최배달",
+            "010-1234-5678",
+            "0",
+            "서울시 잠실역");
 
         RestAssured.given()
-                .when()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberAddRequest)
-                .post("/members")
-                .then();
+            .when()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(memberAddRequest)
+            .post("/members")
+            .then();
         String token = findToken(EMAIL, PASSWORD);
 
         // when
         ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-                .when().delete("/members")
-                .then().log().all()
-                .extract();
+            .given().log().all()
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+            .when().delete("/members")
+            .then().log().all()
+            .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
@@ -187,34 +188,34 @@ class MemberAcceptanceTest extends AcceptanceTest {
     void chargeMileage() {
         // given
         MemberAddRequest memberAddRequest = new MemberAddRequest(
-                "CUSTOMER",
-                EMAIL,
-                PASSWORD,
-                "최배달",
-                "010-1234-5678",
-                "0",
-                "서울시 잠실역");
+            "CUSTOMER",
+            EMAIL,
+            PASSWORD,
+            "최배달",
+            "010-1234-5678",
+            "0",
+            "서울시 잠실역");
         RestAssured.given()
-                .when()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberAddRequest)
-                .post("/members")
-                .then();
+            .when()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(memberAddRequest)
+            .post("/members")
+            .then();
         String token = findToken(EMAIL, PASSWORD);
         MemberMileageChargeRequest memberMileageChargeRequest = new MemberMileageChargeRequest(
-                "5000"
+            "5000"
         );
 
         // when
         final ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-                .body(memberMileageChargeRequest)
-                .when()
-                .patch("members/mileage")
-                .then().log().all()
-                .extract();
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+            .body(memberMileageChargeRequest)
+            .when()
+            .patch("members/mileage")
+            .then().log().all()
+            .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
